consolidate libsafety files (#2808)
* consolidate libsafety files

* cleanup

* even less

* E is silly
