+++ pyproject.toml
@@ -98,6 +98,7 @@ select = [
   "INP001",
 ]
 ignore = [
+  "E302", "E305",  # too restrictive
   "W292",
   "E741",
   "E402",
